requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
dill==0.3.7
    # via wsimod (pyproject.toml)
numpy==1.26.2
    # via pandas
pandas==2.1.3
    # via wsimod (pyproject.toml)
python-dateutil==2.8.2
    # via pandas
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via wsimod (pyproject.toml)
six==1.16.0
    # via python-dateutil
tqdm==4.66.1
    # via wsimod (pyproject.toml)
tzdata==2023.3
    # via pandas
